+++ packages/auth-mobile/test/index.test.tsx
@@ -1,4 +1,4 @@
-describe('@strongly/auth-mobile', () => {
+describe('@strong-js/auth-mobile', () => {
   it('passes', function() {
     expect(true).toBe(true)
   })