+++ packages/auth/src/web/options.ts
@@ -1,7 +1,7 @@
 import { NextAuthOptions } from 'next-auth'
 import Providers from 'next-auth/providers'
 import { PrismaAdapter } from '@next-auth/prisma-adapter'
-import { prisma } from '@strongly/data'
+import { prisma } from '@strong-js/data'
 import { getFullUserByEmail } from '../lib'
 
 export const nextAuthOptions: NextAuthOptions = {