+++ packages/auth/test/index.test.tsx
@@ -1,4 +1,4 @@
-describe('@strongly/auth', () => {
+describe('@strong-js/auth', () => {
   it('passes', function() {
     expect(true).toBe(true)
   })