+++ packages/cli/bin/cli.js
@@ -6,6 +6,6 @@ try {
   require('../dist')
 } catch (e) {
   console.log(
-    '@strongly/cli binary not found, this is expected during first install',
+    '@strong-js/cli binary not found, this is expected during first install',
   )
 }