+++ packages/cli/test/index.test.tsx
@@ -1,4 +1,4 @@
-describe('@strongly/cli', () => {
+describe('@strong-js/cli', () => {
   it('passes', function() {
     expect(true).toBe(true)
   })