+++ packages/crud/test/blah.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import * as ReactDOM from 'react-dom'
 import { Default as Thing } from '../stories/Thing.stories'
 
-describe('@strongly/crud', () => {
+describe('@strong-js/crud', () => {
   it('renders without crashing', () => {
     const div = document.createElement('div')
     ReactDOM.render(<Thing />, div)