+++ apps/web/main/CHANGELOG.md
@@ -1,7 +1,8 @@
-# @strongly/app-main
+# @strong-js/app-main
 
 ## 0.1.1
+
 ### Patch Changes
 
 - Updated dependencies [[`c1a314d`](https://github.com/strongly-labs/strongly/commit/c1a314daff85da271fba691f2619e210dda50f88)]:
-  - @strongly/auth@0.0.2
+  - @strong-js/auth@0.0.2