+++ packages/crud/CHANGELOG.md
@@ -1,8 +1,7 @@
-# @strongly/crud
+# @strong-js/crud
 
 ## 0.0.2
+
 ### Patch Changes
-
-
 
 - [`c1a314d`](https://github.com/strongly-labs/strongly/commit/c1a314daff85da271fba691f2619e210dda50f88) - made public